mysql: let ALTER TABLE run against a table with a primary key
Nearly every table a test suite migrates has an ordinary PRIMARY KEY, and
every ALTER TABLE against one was refused. The durable DDL such a table is
remembered by carries the key inline, and the renderer that writes a table's
MySQL DDL back could not write a PRIMARY KEY at all, so rewriting the table was
refused rather than attempted. It writes the plain inline one now — the shape
the checked slice stores, and only that shape: a key without NOT NULL beside it
is SQLite's rowid alias, which is not a column this ever stores.

Three engine defects were in the way, each found by running the operations.

A column could not be restated under its own name: the duplicate-name check
read the new name against every column including the one being replaced, so
`ALTER COLUMN a TO a TEXT` — changing a type without renaming, which is what
MySQL's MODIFY means — was a duplicate. It now only refuses a name that lands
on another column. Its unknown-column error is typed as NoSuchColumn as well,
so the frontend answers MySQL's 1054 rather than falling through.

Replacing a column dropped its primary key silently. The engine already
refuses a definition that declares one; it refuses replacing a column that
carries one too, which is also what MySQL's MODIFY needs, since MySQL keeps the
key through one.

A schema regenerated during an ALTER lost every column's declared size: a
VARCHAR(8) came back as VARCHAR, and DECIMAL(10,2) as DECIMAL. For SQLite that
reads as a different declaration; for MySQL the length is the type, so no table
with a sized column could be altered at all.

DROP COLUMN on a table with rows in it failed for a fourth reason: the rewrite
writes each stored row back under the schema it is replacing, and the
assignment validator checked those rows against the schema on its way out,
where the column counts cannot agree. Such an insert now says so, and
validation — which is for what a client assigns — skips it.

Tests: cargo test -p turso_mysql_parser -p turso_mysql -p turso_mysql_server -p turso_core
Tests: sqltest sqlite/conformance/turso-sqltests (1585 passed, 2 pre-existing failures)

+++ core/translate/insert.rs
@@ -3776,7 +3776,7 @@ fn emit_update_sqlite_sequence(
         cursor: seq_cursor_id,
         key_reg: r_seq_rowid,
         record_reg,
-        flag: InsertFlags(turso_parser::ast::ResolveType::Replace.bit_value() as u8),
+        flag: InsertFlags(turso_parser::ast::ResolveType::Replace.bit_value() as u16),
         table_name: SQLITE_SEQUENCE_TABLE_NAME.to_string(),
     });
